requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
--find-links https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/torch1.6/index.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu
onnxruntime==1.7.0

dill==0.3.2
cloudpickle==2.0.0
Flask==2.0.3
numpy==1.17.2
Pillow==7.1.2
scikit-image==0.16.2

setuptools<45.0.0
pyinstaller==3.5

scipy==1.4.1
matplotlib<3.2
